Migrate CoinFlipBetGlobal from Kord to JDA
Did some tweaks in UnleashedContext to implement discordInteraction field.
