--- loritta-bot-discord/src/main/kotlin/net/perfectdreams/loritta/cinnamon/discord/interactions/vanilla/economy/bet/coinflipglobal/CoinFlipBetGlobalStartMatchmakingData.kt
@@ -1,11 +1,0 @@
-package net.perfectdreams.loritta.cinnamon.discord.interactions.vanilla.economy.bet.coinflipglobal
-
-import dev.kord.common.entity.Snowflake
-import kotlinx.serialization.Serializable
-import net.perfectdreams.loritta.cinnamon.discord.interactions.components.data.SingleUserComponentData
-
-@Serializable
-data class CoinFlipBetGlobalStartMatchmakingData(
-    override val userId: Snowflake,
-    val quantity: Long
-) : SingleUserComponentData
