+++ loritta-bot-discord/src/main/kotlin/net/perfectdreams/loritta/morenitta/interactions/commands/UnleashedCommandManager.kt
@@ -317,6 +317,7 @@ class UnleashedCommandManager(val loritta: LorittaBot, val languageManager: Lang
         register(EmojiFightCommand(loritta))
         register(RaffleCommand(loritta))
         register(BrokerCommand(loritta))
+        register(BetCommand())
         // April Fools
         // register(CoinFlipBetBugCommand(loritta))
         register(LoriCoolCardsCommand(loritta))
